Moved updateRefreshMenuItem into core package
