+++ core/src/main/res/layout/refresh_action_view.xml
